seventh commit

+++ Banking/src/main/java/com/lcf/app/dao/CustomerDao.java
@@ -6,7 +6,7 @@ import com.lcf.app.beans.CustomerDetails;
 
 public interface CustomerDao {
 	
-	public abstract int newCustomer(CustomerDetails customer);
+	public abstract long newCustomer(CustomerDetails customer);
 	public abstract CustomerDetails getCustomerById(long id);
 
 }
